[fix]: 리스트 선택 오류 수정
